Minor update. Fixes a warning by explicitly using a static_cast

+++ Source/Game/Game/Main.cpp
@@ -28,13 +28,13 @@ class Star {
 			if(this->pos.x > width) {
 				this->pos.x = 0;
 			} else if(this->pos.x < 0) {
-				this->pos.x = width;
+				this->pos.x = static_cast<float>(width);
 			}
 
 			if(this->pos.y > height) {
 				this->pos.y = 0;
 			} else if(this->pos.y < 0) {
-				this->pos.y = height;
+				this->pos.y = static_cast<float>(height);
 			}
 		}
